evaluation: Split evaluator into evaluator+searcher
Evaluator was getting so large it was becoming unwieldy, so this commit
splits it into an evaluator (time control, iterative deepening) and a
searcher (search and heuristics).

The searcher's early return on e.g. a stop() call is handled by a
separate flag to the evaluator's. This clarifies ownership and looks
toward the multi-threading scenario of multiple searchers.

Bench 11349788
