+++ tests/src/test_scoring.cpp
@@ -48,7 +48,7 @@ TEST_CASE("Scoring", "[scoring]")
 
             movegen::ValidMoves moves;
             engine::getAllMoves<movegen::MoveCapture>(*board, moves);
-            s_evaluator.m_moveOrdering.sortMoves(board.value(), moves, 0);
+            s_evaluator.m_searcher.m_moveOrdering.sortMoves(board.value(), moves, 0);
 
             REQUIRE(moves.count() == 3);
             REQUIRE(moves[0].piece() == WhitePawn);
@@ -66,7 +66,7 @@ TEST_CASE("Scoring", "[scoring]")
 
             movegen::ValidMoves moves;
             engine::getAllMoves<movegen::MovePseudoLegal>(*board, moves);
-            s_evaluator.m_moveOrdering.sortMoves(board.value(), moves, 0);
+            s_evaluator.m_searcher.m_moveOrdering.sortMoves(board.value(), moves, 0);
 
             REQUIRE(moves[0].piece() == WhitePawn);
             REQUIRE(moves[1].piece() == WhiteKing);
@@ -77,4 +77,3 @@ TEST_CASE("Scoring", "[scoring]")
         }
     }
 }
-
